% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_images.R
\name{add_images}
\alias{add_images}
\title{Adds images from file to a given PowerPoint presentation (.pptx) file.}
\usage{
add_images(
  files,
  repo_url,
  output_pptx,
  slide_layout_name = NULL,
  base_pptx = NULL
)
}
\arguments{
\item{files}{A list of image file paths to be inserted into the .pptx file.}

\item{repo_url}{The url of the repository where an image is stored.}

\item{output_pptx}{The file path where the modified .pptx file will be saved.}

\item{slide_layout_name}{A character string indicating the name of the PowerPoint slide layout to be used. Default is NULL. If NULL, slide_layout_index is set to 1.}

\item{base_pptx}{The file path to an existing .pptx file that serves as a template. Default is NULL. If NULL, a blank presentation is used.}
}
\description{
Adds images from file to a given PowerPoint presentation (.pptx) file.
}
\examples{
\dontrun{
add_images(
  files = files,
  repo_url = repo_url,
  base_pptx = NULL,
  output_pptx = output_pptx
)
}
}
\keyword{internal}
