% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{This function launches the Shiny application.}
\usage{
app()
}
\value{
This function does not return; it runs the Shiny app until the
session is ended.
}
\description{
This function launches the Shiny application.
}
