% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_layouts.R
\name{extract_layouts}
\alias{extract_layouts}
\title{Extracts layouts as images from a given PowerPoint presentation (.pptx) file.}
\usage{
extract_layouts(base_pptx, output_dir)
}
\arguments{
\item{base_pptx}{The file path to an existing .pptx file that serves as a template.}

\item{output_dir}{The file path where the layout .pngs will be saved.}
}
\description{
Extracts layouts as images from a given PowerPoint presentation (.pptx) file.
}
\examples{
\dontrun{
extract_layouts(base_pptx = base_pptx)
}
}
\keyword{internal}
