% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{install_pptx}
\alias{install_pptx}
\title{Install python-pptx into the reportifyr virtual environment}
\usage{
install_pptx(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical. Print installation messages. Default is \code{TRUE}.}
}
\value{
Invisibly returns \code{TRUE} on success, \code{FALSE} on failure.
}
\description{
Install python-pptx into the reportifyr virtual environment
}
\examples{
\dontrun{
install_pptx()
}
}
