% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_images.R
\name{sync_images}
\alias{sync_images}
\title{Syncs images from file to a given PowerPoint presentation (.pptx) file.}
\usage{
sync_images(input_pptx, output_pptx)
}
\arguments{
\item{input_pptx}{The file path to the input .pptx file.}

\item{output_pptx}{The file path where the modified .pptx file will be saved.}
}
\description{
Syncs images from file to a given PowerPoint presentation (.pptx) file.
}
\examples{
\dontrun{
sync_images(
  input_pptx = input_pptx,
  output_pptx = output_pptx
)
}
}
\keyword{internal}
